+++ server/impl/src/main/java/com/walmartlabs/concord/server/process/pipelines/ResumePipeline.java
@@ -13,6 +13,7 @@ public class ResumePipeline extends Pipeline {
     public ResumePipeline(Injector injector) {
         super(injector,
                 ResumeStateStoringProcessor.class,
+                ResumeDataMergingProcessor.class,
                 RequestDataStoringProcessor.class,
                 StateImportingProcessor.class,
                 EnqueueingResumeProcessor.class);